refactor(chat-api): share Agent-query contract
