+++ apps/chat-api/package.json
@@ -26,7 +26,9 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^2.3.11",
+		"@anthropic-ai/claude-agent-sdk": "0.3.233",
 		"@electric-sql/pglite": "^0.5.3",
+		"@mymemo/agent-query": "workspace:*",
 		"@types/bun": "^1.3.11",
 		"@types/pg": "^8.20.0"
 	}
