+++ packages/agent-query/package.json
@@ -1,0 +1,8 @@
+{
+	"name": "@mymemo/agent-query",
+	"version": "0.1.0",
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts"
+	}
+}
