+++ packages/agent-query/src/index.ts
@@ -1,0 +1,9 @@
+/** The versioned Chat API to AgentCore Runtime request boundary. */
+export interface AgentQueryRequest {
+	version: 1;
+	conversationId: string;
+	conversationEpoch: number;
+	prompt: string;
+	model: string;
+	agentSessionId?: string;
+}
